// src/app/layouts/securitaire/services/models/secure.ts
export class Secure {
    id?: string;
    IdentificationZoneTension?: string;
    ElementViolationDroitHumain?: string;
    ElementCriminaliteUrbainRurale?: string;
    ViolViolenceSexuel?: string;
    NbreGroupeArmeActifRegion?: number;
    InitiavesPaixImplique?: number;
    RelationActeurSecuritSocieteCivile?: string;
    RelationActeurSecuriteMedia?: string;
    RapportGeneraleSecure?: string;
    Province?: string;
    CommunaleUrbain?: string;
    CheferieSecteur?: string;
    Created?: any;
    agentId?: string;
    // ecoImageUrl?: string;
}